feat: run genai tests on multiple models
